- remove template - rename token to currency
